+++ pandas_ml/compat.py
@@ -2,7 +2,11 @@
 
 from distutils.version import LooseVersion
 
+import sklearn
+_SKLEARN_ge_017 = sklearn.__version__ >= LooseVersion('0.17.0')
 
-def _SKLEARN_ge_017():
-    import sklearn
-    return sklearn.__version__ >= LooseVersion('0.17.0')
+try:
+    import imblearn                 # noqa
+    _IMBLEARN_INSTALLED = True
+except ImportError:
+    _IMBLEARN_INSTALLED = False