+++ pandas_ml/imbaccessors/__init__.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python
+
+from pandas_ml.imbaccessors.base import ImbalanceMethods  # noqa