+++ scripts/build_travis.sh
@@ -22,3 +22,7 @@ python -m pip install graphviz
 python -m pip install flake8
 python -m pip install xgboost
 python -m pip install enum34
+
+if [ "$IMBALANCE" ]; then
+    python -m pip install imbalanced-learn
+fi