+++ doc/source/whatsnew.rst
@@ -8,7 +8,8 @@ v0.4.0
 Enhancement
 ^^^^^^^^^^^
 
-- Support scikit-learn v0.17.x
+- Support scikit-learn v0.17.x.
+- Support imbalanced-learn via ``.imbalance`` accessor.
 
 Bug Fix
 ^^^^^^^